requirements.txt
anyio==3.4.0
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
distlib==0.3.4
fastapi==0.70.1
filelock==3.4.0
greenlet==1.1.2
h11==0.12.0
httptools==0.3.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
platformdirs==2.4.0
pluggy==1.0.0
py==1.11.0
pydantic==1.8.2
pyparsing==3.0.6
pytest==6.2.5
python-dotenv==0.19.2
PyYAML==6.0
requests==2.26.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.28
starlette==0.16.0
toml==0.10.2
tox==3.24.4
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
virtualenv==20.10.0
watchgod==0.7
websockets==10.1

